#53: modularized 2 (not tested)
