+++ terraform/aws/glue_job/src/modules/jdbc.py
@@ -1,6 +1,15 @@
 from pyspark.sql import SparkSession, DataFrame
 
 
+def get_jdbc_options(jdbc_url: str, secrets: dict) -> dict:
+    return {
+        "url": jdbc_url,
+        "user": secrets["username"],
+        "password": secrets["password"],
+        "driver": "org.postgresql.Driver",
+    }
+
+
 def fetch_table_names(spark: SparkSession, jdbc_options: dict) -> list[str]:
     query = """
         (SELECT table_name
